Be able to inject secret using environment-variables
!!! This is might be a BREAKING CHANGE for some users !!!

The HaliteEncryptor and DefuseEncryptor used to accept the filepath to
the keyfile as an argument in the constructor.  This has been modified
so they accept the actual secret as an argument.

A new secretFactory was created, which will fetch the secret from the
existing file using the existing `secret_directory_path` and classname
of the used encryptor.

An option (`enable_secret_generation`) was added to disable the automatic
generation of keyfiles when they do not exist yet to allow users to
avoid accidentally creating new secretfiles.

An option (`secret`) was added to directly inject the secret, allowing
users to environment variables and environment variable processors
for setting the used secret.

This might be a breaking change for users if they have extended our
base-encryptors or use our base-encryptors outside of the library by
instantiating the classes themselves.

If a user was using the library as expected, there will be no breaking
change as the library behaves the same if only options were used.